+++ tests/MessageMode/test_Cloud.py
@@ -48,4 +48,4 @@ class TestCloud(object):
     def test_sdk_version(self):
         cloud = Cloud(None, send_host = '127.0.0.1', send_port = 9999)
 
-        assert cloud.version == '0.7.5'
+        assert cloud.version == '0.7.6'